airlockBooth logic fix

+++ Projects/Common/GKProcessor/DescriptorsDatabase/Descriptors/PimDescriptor.cs
@@ -15,8 +15,8 @@ namespace GKProcessor
 
 		public override void Build()
 		{
-			DeviceType = BytesHelper.ShortToBytes((ushort)0x107);
-			SetAddress((ushort)0);
+			DeviceType = BytesHelper.ShortToBytes(0x107);
+			SetAddress(0);
 			SetFormulaBytes();
 		}
